megatron: implement sampler based on megatron sampler data: remove distributed.py
